requirements.txt
asgiref==3.4.1
Django==3.2.5
django-digest==1.13
django-environ==0.4.5
django-reset-migrations==0.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
FormEncode==1.3.1
linecache2==1.0.0
lxml==4.6.3
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
pyxform==1.5.1
six==1.16.0
sqlparse==0.4.1
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
xlrd==1.2.0
